+++ src/tools/rbd_wnbd/wnbd_handler.h
@@ -175,15 +175,15 @@ private:
   static void PersistResIn(
     PWNBD_DISK Disk,
     UINT64 RequestHandle,
-    UINT16 ServiceAction,
+    UINT8 ServiceAction,
     PVOID Buffer,
     UINT32 AllocationLength);
   static void PersistResOut(
     PWNBD_DISK Disk,
     UINT64 RequestHandle,
-    UINT16 ServiceAction,
-    UINT16 Scope,
-    UINT16 Type,
+    UINT8 ServiceAction,
+    UINT8 Scope,
+    UINT8 Type,
     PVOID Buffer,
     UINT32 ParameterListLength);
